Creates folder stand_up_bot and moves code related to
