+++ stand_up_bot/greetings.csv
